Correctly link to the ocoms depedency.
When correctly installed hcoll will pull the ocoms library
automatically. However, as this library is installed with an rpath if
the installation directory moved to a different location the ocoms
library will not be automatically pulled during link time. By adding it
explicitly we force the linker to look for it (will be found in the same
directory as hcoll) and use it during link.

Signed-off-by: George Bosilca <gbosilca@nvidia.com>